Actually fixed superscript and subscripts
Closes #19;

NB: this changes the order in which primes and superscript are drawn (I
think for the better). Previous, super scripts were first followed by
primes. Now, primes are first.

The old behavior can be recovered by manually passing a formatted prime
symbol (\prime, or \tprime, or whatever) in the formatted superscript
argument to a metavar.
E.g.
\txmetavar{}{\tprime1}{0}